requirements.txt
appdirs==1.4.4
black==20.8b1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
future==0.18.2
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
line-bot-sdk==1.17.0
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
